build: Moved logger into it's own directory
Created the logger as it's own project and will set it up there so I never look at it again. When
it's setup in it's own directory, I'll know that it's working and the checks are passing. This is
just for peace of mind, and hopefully someone else may find it useful. I think it's extensible and
ready for action and here's the proof with checks.